修改支付金额时的元转分的处理，必须整形

+++ admin/api/order.py
@@ -192,7 +192,7 @@ def random_digit(length=10):
 def create_order(card_id, amount, openid, count):
     customer = Customer.query.filter_by(openid=openid).first()
     order_id = random_digit(16)
-    pay_amount = int(amount * 0.99)
+    pay_amount = amount * 0.99
     order = Order(order_id=order_id, card_id=card_id, customer_id=customer.openid, face_amount=amount,
                   card_count=count, pay_amount=pay_amount, order_type=1, paid=False)
     cache_order(order)
